Add adaptive TP_ALLOC_CACHED instruction
Summary:
This adds an adaptive `TP_ALLOC_CACHED` instruction. We cache the value in the class loader and patch it with an index to load from. We'll end up using the cache for many other instructions later.

If we clear the caches during tests then we clear the list and bump a base index, so that any previous lookups will fail and fallback to the non-opt versions.

Reviewed By: martindemello

Differential Revision: D75182991

fbshipit-source-id: a978b04a53b02c98c7da5dfbf6b53bb14a2669ef