Add adaptive interpreter support for INVOKE_FUNCTION
Summary: This adds adaptive support for `INVOKE_FUNCTION`. This comes in two forms, the first is another object cache similar to `TP_CACHE` where we store the function directly. The 2nd is when we have the indirect form - this happens with a mutable container. Here we burn the pointer directly into the cached value as it uses the JIT watcher infrastructure which will keep the pointer alive.

Reviewed By: martindemello

Differential Revision: D75259729

fbshipit-source-id: ea750324eed9135c482344660dec05c35dfc855a