Implement specialization for BUILD_CHECKED_MAP
Summary:
Adds specialization for the `BUILD_CHECKED_MAP` opcode - nearly identical to the list variation, again we just need to remember the specific list type so we use the class loader cache.

We also start getting up to opcode 229 in this diff, which is apparently used in `test_code` as an invalid opcode. So our assignment script is updated to skip over this.

Reviewed By: martindemello

Differential Revision: D75830239

fbshipit-source-id: 1b0ae265e02fd33a72d1381b704b2dd4aa7cd733