Prefer to use AyncLazyValue from _asyncio
Summary: Currently we are potentially getting 2 copies of `AsyncLazyValue` - the Cinderized version and the version from igasyncio. This makes it so that we always try and get it from `_asyncio` first which is the replaced IG version. This can potentially cause some inefficiences where the async io module is looking for the specific AsyncLazyValue type.

Reviewed By: SonicField

Differential Revision: D76313131

fbshipit-source-id: e864f3a4f50932b9735e1cda910bd7f67033fd82