Add something like shadowed instances
Summary:
In 3.10.cinder we track whether or not an instance ever has an attribute instance which shadows a method.

In 3.12 we can get close to this by checking to see if an instance attribute exists in the shared keys. This requires an extra check when doing the lookup to make sure we're looking in the dictionary keys are the version that we cached against. The dictionary key versions are interpreter unique so we don't have to check the keys object themselves.

Reviewed By: martindemello

Differential Revision: D76087393

fbshipit-source-id: f08db39d561eda7cc11681d1d14527e52685d51b