Support managed dictionaries in inline caches
Summary:
Currently we get no load attr caches for user-defined types because they'll typically have a managed dict and those have a `tp_dictoffset` set to -1, this causes us to not be able to initialize the split attribute mutator.

In 3.12 only objects with managed dictionaries use the cached keys, so the split mutator is updated for this condition. We need to use `_PyObject_DictOrValuesPointer` to get the dictionary. But in the common case no dictionary will ever be created - we'll only have the values portion of the object and a dictionary will only be materialized if something accesses `__dict__` directly. So we need to handle the case when we have values and get/store the value from there.

Another change in 3.12 is that objects don't need to have the attributes stored in the same order in order to keep the cached keys optimization working. Instead the order is tracked on a per-object basis, so if we're doing the first insert we need to update the order.

Reviewed By: martindemello

Differential Revision: D76087394

fbshipit-source-id: ff746725398a50e81e4f65cf431637794ab1ae71