Reduce stack space used in dev builds of the CinderX interpreter loop
Summary:
A number of tests have started failing with stack overflows in dev builds that have CinderX enabled.

This is likely just because the interpreter loop can't be as well optimized. Let's not explicitly inline some functions in the interpreter loop and use less macros and let the compiler figure out the best path forward.

Reviewed By: alexmalyshev, jbower-fb

Differential Revision: D76443007

fbshipit-source-id: 076d083398d257f99adf817915c27ea864c00924