Enable typed method def only on 3.10
Summary: We only use the type method defs on 3.10, so this just ifdef's out the code that has stubs in it and gets rid of the stubs. We do still install `__typed_signature__` on some types so those functions need to stick around and return None for the static compiler to reflect on them and do nothing.

Reviewed By: alexmalyshev

Differential Revision: D76548722

fbshipit-source-id: 21d28860ee180b658650d08e16e89c372c52d20b