Extract frame assembly
Summary: This extracts the code related to Python frame handling into a stand alone class separate from the rest of gen_asm. This stuff is pretty stand alone as-is and so simplifying gen_asm will be nice. But we'll also get one more variation of this when we start supporting more lightweight frames in the JIT generated code.

Reviewed By: jbower-fb

Differential Revision: D76908316

fbshipit-source-id: 400c9387ea21b94cdd577f422fb77416d3088b84