Use static allocated thread space
Summary:
This switches to using stack based inline frame allocation - while this is much more involved than our 3.10 shadow frames it's a lot more efficient than our current 3.12 calls out to a helper.

We still need to store the function somewhere like in the previous diff so we can reconstitute it so for simplicity it moves to before the `_PyInterpreterFrame`.

When a frame is unwound and the frame object outlives the object we need to re-allocate a new `_PyInterpreterFrame` on the slab of frames and then update the existing frame to point at this one. This means we no longer need to specifically allocate an extra slot for the frame.

We also now immortalize the helper object so that we don't have to do any ref count operations on it in the prologue.

There's a little bit of magic in here in that getting to the current thread state is difficult in 3.12. It's stored in a C thread local and `_Py_tss_tstate` is not exported but even if it was I'm not sure there's a much cleaner way to get the address. If it was exported we might be able to go digging around the loaders data structures similar to `__tls_get_addr` and resolve it to a specific offset and burn that in. But that seems less safe than just knowing what the precise assembly does in the optimal case.

References:
https://chao-tic.github.io/blog/2018/12/25/tls#the-internals-of-elf-tls
https://maskray.me/blog/2021-02-14-all-about-thread-local-storage

Reviewed By: jbower-fb

Differential Revision: D76873985

fbshipit-source-id: f4aecd1808e0ae61c0d4290cad7c9e5926563003