Refactor - add RegisterPreserver class
Summary: This is just a small refactoring to extract our register preservation logic into a stand-alone helper class. In the future this will let us create an instance which has the saved registers and pass it down so that we'll be able to conditionally preserve registers based upon the type of frame we're linking.

Reviewed By: SonicField

Differential Revision: D76908163

fbshipit-source-id: 55d9315db1e0b52b00a73ec1f12b41de69d73102