Add explicit PYTHONJITALL and -X jit-all options
Summary:
PYTHONJIT=1 shouldn't mean "JIT-compile all the things ASAP".  That's what we
generally refer to as "JitAll", and so we should have an option for that.

At the same time, stop warning when JIT options are passed but the JIT is
disabled.  They're not particularly helpful and lead to a lot of repetition.

PYTHONJIT=1 currently implies PYTHONJITALL=1.  This will be changed in a
future diff.

Reviewed By: DinoV

Differential Revision: D73401848

fbshipit-source-id: 72821b19bdd900d189fa208b0d09e80d705d0659