Normalize opcode generation
Summary:
Our opcode generation scripts are similarly setup in an ad-hoc manner. This moves things to directories. It also brings over 3.10's `generate_opcode_h.py‎` which we currently fork and include in CinderX for 3.12.

This is going to undergo some more changes in a future diff, but for now this gets us to a cleaner spot.

Reviewed By: martindemello

Differential Revision: D77264702

fbshipit-source-id: 1e09146e63862096fdfac97f3ba44766c61aa0cf