Do not assume there will be last uses in isVectorCallIfIsInstance
Summary: If there are no last uses for an instruction there will be no entry at all in the last_uses map, causing `::at()` to fail with an exception. I'm not sure why, but this seems to happen in 3.12 rather than 3.10 and bubbles up quietly terminating compilation of the current function.

Reviewed By: alexmalyshev

Differential Revision: D78051052

fbshipit-source-id: ac1afc0cece473b0ba803c799917aabc03e26762