Fix argument comment in cinderx/RuntimeTests/bytecode_test.cpp +3
Summary:
Argument comments ensure that the correct values are assigned to the correct arguments.

This diff is a manual fix (assisted by the script in D78190897) of one or more argument comments in cinderx/RuntimeTests/bytecode_test.cpp. Once existing violations are drawn down we will make mismatches between argument comments and argument names an error.

Reviewed By: dtolnay

Differential Revision: D78191313

fbshipit-source-id: 568626d6abadb1d6577b5f100ab536f60db9c621