Move JIT data into generator objects
Summary:
In 3.10 generators were fixed size, however as they now contain `_PyInterpreterFrame` (rather than pointing to one), they are "variable". So, stuffing some extra data on the end for JIT isn't a big deal. Unfortunately, we still need to add an additional field to point to the JIT data as there's no way to derive the size of this directly from the CPython generator data. We could eliminate this if the generators were truly variable size, but actually they aren't: https://github.com/python/cpython/issues/126596. On the upside, this means very little needs to change in existing code.

Note in a Swayze run, this actually makes things *slower* by as much as ~0.3%: https://fburl.com/instagram/sa68hau3. This is reversed once we add a free-list for JIT generators later, but I'm actually not sure why things are worse now. The allocation of JIT data was already very cheap because we have a free-list just for that, but we still had to allocate the generator objects anyway and I don't see why making them a little bigger would be such a performance hit.

I also actually ran an experiment using the existing free list for JIT data and a new free-list just for the existing generator data. This was neutral: https://fburl.com/instagram/qfqan16f. 🤷‍♂️

Reviewed By: DinoV

Differential Revision: D77571142

fbshipit-source-id: 4b66f111f54db2c294a46e3a26c8c6864555a89b