Cache functions and static methods
Summary: In 3.12 the compile is pretty aggressive about not emitting the equivalent of `LOAD_METHOD`/`CALL_METHOD` on things which are imported from modules. This means doing a call to a function on a type can end up going through `LoadTypeAttr*` instead of `LoadTypeMethod*` and we don't get caching when it's a function. This allows us to cache functions and staitc methods as we know what the results of them are.

Reviewed By: martindemello

Differential Revision: D78347891

fbshipit-source-id: f87d01ac7502fddec8ba445aaaa25f2d2788a140