Set ENABLE_* flags in setup.py
Summary:
So turns out CMake doesn't actually support arbitrary expressions in set()
commands.  Doing something like

  set(A (B AND C))

will set the value of A to the string "(B AND C)".  The AND operator is only
recognized within the context of an if() command.  Move this logic to Python
where things make sense.  CMake doesn't make sense.

Reviewed By: mpage

Differential Revision: D78425224

fbshipit-source-id: 04accfe18cfde106228d7a2e33a74105ca98e3f7