Fix checked dict 3.14 compatilility
Summary: This removes a few lingering spots where we're crossing the streams between CPython's dicts and the checked dict implementation. A big chunk of this is not using CPython's `_PyDict_HasSplitTable` - we never have split checked dictionaries. There's more that can be done here to remove split support but this does the bare minimum for now.

Reviewed By: alexmalyshev

Differential Revision: D77566447

fbshipit-source-id: 92ea79dc7fc26db7dfb272729fa183c5e31b0080