Make list free-list be CinderX specific
Summary: We shouldn't really be mixing our checked lists w/ the normal list free list in the first place, but we can get away with it because they have the same layout. But rather than continuing down that path this just adds our own free list which is what we did for checked dict in 3.12 when the object layouts diverged. The 3.14 free list mechanism is unfortunately not extensible to our own types because some space needs to be allocated in the runtime.

Reviewed By: alexmalyshev

Differential Revision: D77566448

fbshipit-source-id: 777266f12846c41810d9c7f534775e4df684bfa5