Put eval hook under ifdef
Summary:
We want to switch away from using the eval hook in 3.14+. This diff adds a new feature flag which does so. We have a couple of issues which prevent this working on 3.14:
1) we will conflict with the perf trampoline support which we'll want to patch 3.14 for and backport a fix to 3.15
2) Several quickenings will fail/be disabled when the hook is in place.

We can address 1 early on in meta-Python (it involves a change to the runtime which low-level tools like Strobelight may care about the offsets in).
We can address 2 in 3.15 with a plan to let the interpreter loop know the eval hook is safe to quicken against.

Reviewed By: alexmalyshev

Differential Revision: D77550892

fbshipit-source-id: 169304f78b3ffe48913ed0aba7f8d461c09a268b