Add cinderx/python.h, handle atomic symbol collisions
Summary:
Adding a new header to wrap Python.h and to force the right order between atomic
and stdatomic.h.  This is only really an issue with 3.12 and later.  3.10.cinder
seems fine, in fact it gets upset if we include atomic headers before
internal/pycore_runtime.h so we have to avoid that.

All usage of Python.h should be replaced with cinderx/python.h now.
There's an exception carved out for cinderx/PythonBin to keep it simple.

Reviewed By: DinoV, mpage

Differential Revision: D78585773

fbshipit-source-id: 1d10410c43e1437413cee8c8692d9c410c34f5f6