Forward awaits on JIT co-routine wrappers to the co-routine
Summary:
S543823 is exposing a lack of awaiter pointers in some cases.

`igasyncio` is not aware of the JIT generators and is causing a Python function that produces a co-routine wrapper instead of returning the raw co-routine. co-routine wrappers do not forward the async awaiter to their underlying co-routine meaning we lose the chian.

We're still investigating the right fix. Making `igasyncio` depend on CinderX has some drawbacks and is somewhat complicated. There may be other options though that we'll investigate.

For now this causes JIT generator coro wrappers to forward the async set to the underlying coro.

Reviewed By: yulong-zhang-1

Differential Revision: D79020862

fbshipit-source-id: dce392b3c9dad9b90d3ff2c840b6688de2b10b25