Replace use of TODO(Txxxx) with TASK(Txxxx)
Summary:
TODO has an implication that the developer intends to change something make
before they land their commit, or that there is a known bug in the code.  This
is not quite how our team uses it, we generally use it to point to future
improvements that can be made to the codebase.  We like tying code together with
relevant tasks, and so TODOs don't seem like the appropriate tool.

Switching to "TASK" because that's another four letter word so we won't need to
reflow code.

Note: This diff only goes after TODOs with tagged tasks.  We still have
TODO(unixname) comments lying around and those should definitely be cleaned up.

Reviewed By: DinoV, yoney

Differential Revision: D78915306

fbshipit-source-id: ebe8c21f0be1db1dd95b707203d2d0e452beed62