Fix crash when frame w/ no locals is reified
Summary:
This fixes a crash when we reify a frame that has no locals and no stack - in the case of S544761 we're hitting this with `lambda: None` which is probably one of the few ways to hit this. Presumably some other thread in the process is getting all of the stack frames for all of the threads causing the reification to happen.

Our stack frame is laid out [like this](https://www.internalfb.com/code/fbsource/[710c2bf3b9ef]/fbcode/cinderx/Jit/codegen/gen_asm.cpp?lines=875). Note this is following "the stack grows down", so lower things on this list are actually at lower memory addresses.

In this case our `_PyInterpreterFrame` starts at rbp-0x48 and our `FrameHeader` is in front of that at rbp-0x50. Our "spilled values" space should be lower than the `_PyInterpreterFrame` but we're just blindly using -0x8 which before the lightweight frames was always the "first" spilled space. This -0x08 is basically light negative indexing into the `_PyInterpreterFrame` struct.

And the [`_PyInterpreterFrame`](https://www.internalfb.com/code/fbsource/[017fe48c29bc8e229de39eea940735dddd8fb8e8]/third-party/python/3.12/Include/internal/pycore_frame.h?lines=51) struct includes space for "localsplus" - this is local variables and stack space.

So when we have stack space we're just blindingly updating locals + stack to contain this value. That's almost always perfectly fine because this is locals and then stack, our stack pointer is at 0 as we unwind, and so we wouldn't decref any of the stack values. Getting a frame with locals but no stack is probably a little more difficult than a frame with no locals and no stack.

But when we have no locals and no stack we're actually updating `stacktop`, `return_offset`, and `owner` as they fit in one word. That causes us to try and clear a large number of items "in the frame" and not surprisingly we crash.

The move to -0x8 is questionable even not when using light weight frames. Has `gen_asm` actually allocated any scratch space? And even if it has, who owns it? So now we make the frame space allocation explicit.

Reviewed By: diansheng92, yulong-zhang-1

Differential Revision: D79109552

fbshipit-source-id: 01152e1d5c6e1fa473e6cd92a210ca41ee5b25d5