Unify testing
Summary:
This gets rid of the bifurcated testing that we have between 3.10 and 3.12 and uses a singular code path for all of testing. This will let us on-board some 3.14 tests next.

At this point most of our .json files which track failures are all failure-free. The one notable exception here is there's a single failure for `dev-nosan-refleak-failures`. This gets rid of all of those files and drops down a single file of test-failures for that mode. To unify things we now pass a list of failures off to the test runner and let it ignore the failed tests and report a failure if a failed test doesn't fail. Note we don't need this functionality for 3.10 because we have no failing tests there so it just doesn't get it. Similar to before we output an updated list of tests that are failing.

One other thing that needs to be normalized is supporting `-R` on the 3.10 test running instead of needing to pass it to the CPython script - this just lets us get rid of one more small delta.

We also split out our executable variants from our test variants. So we now get an executable for jit-all, and we actually have one for jit-auto as well now as it doesn't bring along the tests. This lets us define the supported set of tests at the language version level. Ideally we'd have these all be the same for all supported platforms but we still have some more work to get there. But this continues bringing us a single unified place to see the differences between our version support.

The end result is that all of the same commands should be available (e.g. `buck run :python-tests3.12`, `buck run :ctr3.12`, etc...). The 3.12 tests end up with slightly different names in CI because the order of version / mode was slightly different than 3.10.

Reviewed By: alexmalyshev

Differential Revision: D78769885

fbshipit-source-id: 53458437a567b8bfb3caa3c8b1baf0043826de68