Fix multi-threaded compile issues
Summary:
In 3.12 there are more cases where where Python libraries try to get thread state as this is needed to get the current interpreter. This is incompatible with multi-threaded compile where we don't have an initialized thread-state, and my attempts to initialize it have not gone well. Multi-threaded compile probably needs to die and be replaced with something better in a Free Threaded world. But for now this fixes two places which break in multi-threaded compile:
* We now skip the `simplifyLoadAttrInstanceReceiver` optimization if the type doesn't already have a version tag. Fortunately this seems to be very infrequent, and we can still apply it in non-multithreaded mode.
* We now defer a call to `_PyClassLoader_HasPrimitiveArgs` in `Runtime::findFunctionEntryCache`.

Reviewed By: alexmalyshev

Differential Revision: D78602783

fbshipit-source-id: b2d261530bc5671977de9b722d160a00067ddca2