Add custom transform for eval frame checks in specialize
Summary:
The specializer checks to see if the eval frame hook is present and if so refuses to specialize functions. This causes some test failures that expect specialization.

This adds support to have custom transforms which are defined per-function in the borrow script and adds one which transforms the eval frame check to account for our eval frame hook.

Reviewed By: alexmalyshev

Differential Revision: D79363228

fbshipit-source-id: 31bf87884ba19b2ed0bc7cfdef16c88091c59764