Make opcodes eval frame hook aware
Summary: There's a number of opcodes that want to check and see if we have the frame evaluation hook installed. We want to modify these to check and see if the frame evaulation API is our function. This copies the bytecodes from 3.14 and replaces the eval_frame check with a `IS_PEP523_HOOKED` check.

Reviewed By: alexmalyshev

Differential Revision: D79363230

fbshipit-source-id: 088c5ec5facc83c843fed8eeecd2c9e3d0d9f5ea