Allow calling SlabArena::mlock() more than once
Summary:
This happens when we run JIT tests with refleak detection, and we end up
crashing on the JIT_CHECKs here.

It's safe to call mlock() on a region of memory more than once.  We technically
don't need to track the count because nothing actually calls
SlabArena::munlock() yet, but putting it in plus JIT_LOGs defensively in case
that changes in the future.

Reviewed By: yoney

Differential Revision: D79661397

fbshipit-source-id: f30298b5df73bf3439c201d088f048000947b3f2