Add "free-list" for JIT generator objects
Summary:
Redo of D77572569. The original had a bug where upon shutdown the module state was free'd but we still go down the `jit_dealloc_gen()` path which tries to make use of it.

We now swap back the original dealloc methods for the system generators on shutdown. We also move the free list data to the module state and bump the module state's refcount so we don't free memory from underneath outstanding JIT generators.

Reviewed By: alexmalyshev

Differential Revision: D78516909

fbshipit-source-id: 5ff870968d90873bedd1b4a34edae132133f4ee9