Support running test_jit_disable.py functions multiple times
Summary:
The test functions here need to be explicitly uncompiled (and in one case,
unsuppressed) otherwise their compiled state will stick around on the next test
run.  This happens when running the refleak tests.

`force_uncompile` needs to get stronger about forcefully dropping all
compilation state for a function.  Its current behavior is effectively just a
deopt, if we create another function with the same code object we'll reopt it
immediately upon calling lazy_compile(), which is not what we want here.

Reviewed By: czardoz

Differential Revision: D79690872

fbshipit-source-id: 52f14546cf67cb0f7ac6f3a9bd1b8fe55f35d256