Re-enable load method elimination for 3.12
Summary:
Re-enables the load method elimination optimization in multi-threaded compile that's currently disabled in 3.12.

We maintain our own dictionary of core built-in types which we can apply this to. To make this thread safe in 3.12 we use an `unordered_map` for the type->dict mapping as we can't safely use a map that's keyed off non-unicode keys without potentially hitting the interpreter state. We then use a normal Python dict w/ only unicode keys for the members which is thread-safe w/o the GIL held.

If a type isn't one of the built-in types but is immutable, has no meta-class and has no dictionary we also allow the optimization to be applied (although this hits very few types in the test cases - mostly just something in the lru cache).

The `unordered_map` is our first usage of a C++ type in `ModuleState` which has initialization that's something other than zero-initialization. Therefore we now run the in place constructor on the memory that the Python runtime has allocated for us to store the ModuleState.

Reviewed By: mpage

Differential Revision: D78361677

fbshipit-source-id: ec3280d28dc434bb2b2e4f1a81156030e4aab3ac