Extract code-runtime into it's own library
Summary: We have some dependencies we need to sort out to allow the frame target to be able to access the things it needs to do to update the frame. This is the first part - it extracts the `CodeRuntime` into its own module as it's really just some simple data. That decouples it from the test of the `:runtime` which has a lot of state and dependencies.

Reviewed By: alexmalyshev

Differential Revision: D78193675

fbshipit-source-id: 7ef026b34e61bd2910a782ff6012799ab80a97bc