Track CodeRuntime in CompiledFunction
Summary: Ultimately these refactoringings are letting us get back to the `CodeRuntime` for a given function object. But in order to do that we need to track the CodeRuntime along with the `CodeRuntime`. This is pretty easy to do as `gen_asm` produces both - it just never connects them. So now we connect them.

Reviewed By: alexmalyshev

Differential Revision: D78193676

fbshipit-source-id: 51b83bf7b028e21e40e51aafb48b893e00551660