Update prev instr on frame reification
Summary: Finally we can start putting this together - now that we can get at the necessary data we can re-construct the current instruction pointer when we attempt to update the frame information. This is resurrecting a lot of the existing unitState machinery which exists for 3.10 but it has lot of minor changes to accomodate the 3.12+ machinery.

Reviewed By: alexmalyshev

Differential Revision: D78193674

fbshipit-source-id: 91e903401805009cee326ae938c95e92efe0912b