Extract frame header into it's own library
Summary: The previous diff had some exceptions for generators where we're not yet using the reifier and the light frames. In order to do this with generators we need to be able to embed the `FrameHeader` into the gen data footer. This extracts the frame header out into it's own library instead of having it embedded in `:frame`.

Reviewed By: mpage

Differential Revision: D78593316

fbshipit-source-id: 8ab9664024b0323c6ac4799d3c5c1a8e68c5c928