Add lightweight generator frames
Summary:
After moving a lot of stuff around this finally brings us generator frames which get the same benefits of the lightweight JIT frames. We no longer need to initialize the full structure and we support not updating prev instr with the generator instead deferring that to when the frame is accessed.

We embed the FrameHeader within the generator footer and our access to it is updated.

Reviewed By: alexmalyshev

Differential Revision: D78675675

fbshipit-source-id: c844bf9d86b53ba827e5cddb9da9d3eccfe1aad4