Extract gen-data-footer into own buck target
Summary: Next we need to extract `gen-data-footer` - again this is largely just data - although it also includes how to get at it from a JIT generator object. We also will need this in `:frame` so that we'll be able to access the footer and get the frame header for a generator which needs to be stored at a different location.

Reviewed By: alexmalyshev

Differential Revision: D78593315

fbshipit-source-id: bdc6edd534542ce47204039e7065c9eee0d3af6c