Initialize the JIT for PYTHONJIT=1, but don't set PYTHONJITALL
Summary:
This is a cut down version of D73401869, which was reverted because it hit
issues with refleak tests.  This diff doesn't enable the JIT for the refleak
tests.

Allows initializing the JIT without assuming any particular mode of operation.
Users can call out to manual operations like `cinderx.jit.force_compile()` and
have them work. But to automatically compile code users still need to set
`PYTHONJITALL`, `PYTHONJITAUTO`, or `PYTHONJITLISTFILE`.

Trying to clarify how some of the logic works now: `shouldCompile` is now renamed
to `shouldScheduleCompile`. It designates whether the JIT decides it should compile
something, based on the environment variables listed above.  It is **not** called when
the user calls `cinderx.jit.lazy_compile()` or `cinderx.jit.force_compile()`,
those explicitly bypass it.

Reviewed By: martindemello

Differential Revision: D79830074

fbshipit-source-id: b112d24b734b6591499a7a6deeed8c56fac1527f