Release CodeRuntime references in CompiledFunction dtor
Summary:
If the compiled function is going away, then so should the references to its
CompilationKey.  There's no more need to keep around the code, builtins, and
globals references.

Currently this only kicks in when `cinderx.jit.force_uncompile` is called on a
Python function.  Otherwise all outstanding CompiledFunction objects and
CodeRuntime references will be collected when the CinderX module is finalized.

Reviewed By: DinoV

Differential Revision: D80044795

fbshipit-source-id: 82152edfe36f55b7771c0ecd9b4ec9d4c884c2f3