Don't assert that functions exist after preloading
Summary:
Preloading does all sorts of weird behaviors.  When we try to compile a Python
function there is the chance that we blow it away and end up only compiling the
functions **it** calls instead.  This is a weird place to end up in, but it's
not quite an error, so return `PYJIT_RESULT_OK` for now.

Reviewed By: mpage

Differential Revision: D80226542

fbshipit-source-id: 28a0de1d6f9fe32211038508121f86fae09d14b4