Add more asserts on types flowing into GlobalCache
Summary:
We're depending on implicit casts between PyObject* to BorrowedRef<PyDictObject>
and Ref<PyUnicodeObject> in a few places.  Make this more robust.

Similarly, assert that key names are interned during construction of GlobalCache
values, instead of spraying around asserts in multiple places.  The only other
place that needs the check is notifyDictUpdate(), as thats a value being given
to us by the dict watcher handler.

Reviewed By: yoney

Differential Revision: D80299756

fbshipit-source-id: 4048e521aa4b50ab0844a65698eb1cbca0d8d1f0