Make PyImport_IsLazyImportsEnabled only base it's decision on if lazy imports are enabled [3.10]
Summary:
There's an API to check if lazy imports are enabled, but it consults the module that's executing the closest instead of returning if they're active at all. This is just wonky and definitely not how the API is documented...

Do these callers expect that this returns true if and only if the nearest executing module isn't on the excluding list?

https://www.internalfb.com/code/fbsource/[3cc96883674dbeffda436cd8e4e9b04c5aab5d8a]/fbcode/python/build/e2e/lib/basic_info.py?lines=33
https://www.internalfb.com/code/fbsource/[3cc96883674dbeffda436cd8e4e9b04c5aab5d8a]/fbcode/minimal_viable_ai/mvai_cli/run.py?lines=219
https://www.internalfb.com/code/fbsource/[3cc96883674dbeffda436cd8e4e9b04c5aab5d8a]/fbcode/frl/ctrl/pte/trainer/py_conf/configs/pipelines/f5/sanity_check.py?lines=48-49%2C53-54

Reviewed By: itamaro

Differential Revision: D80203460

fbshipit-source-id: cb803ac5685a62bd9795b0d1b058ab44afc3945b