Simplify TO_BOOL/UnaryNot conversion logic
Summary:
As alexmalyshev points out this can be done without explicit branching by using `PrimitiveBoxBool`.

However, there is a catch for `TO_BOOL` which is the output type of `IsTruthy` is actually `CInt32` and `PrimitiveBoxBool` wants a `CBool` input. To address this I've used a `IntConvert` and extended this to support `CBool` as an output.

Another approach I explored is making the output type of `IsTruthy` into `CBool`. This is probably doable with enough work, but it's non-trivial right now because a few things, particularly simplification passes, assume `IsTruthy` is of type `CInt32`.

Reviewed By: alexmalyshev

Differential Revision: D80776121

fbshipit-source-id: 640be202c6a4d2004e1c31fa7ab27a60bc0a1853