Unconditionally define Parallel GC Python functions
Summary:
The conditional compilation makes it too hard to reason about and annoying to
work with.  Have the native extension always define the functions, but have them
be stubbed out when the support isn't enabled.

enable_parallel_gc() will now raise a RuntimeError if it's called to enable the
Parallel GC but it's not supported in the build.  The other functions don't have
to raise an exception, they're fine with returning None.

Reviewed By: DinoV

Differential Revision: D80729662

fbshipit-source-id: 03552391b4d110f9c5d2debf1624957a59035cd0