Add 3.14 compiler infra
Summary: Adds basic infra for the 3.14 compiler. This includes the various subclasses that we need to start filling in as well as the opcode definitions. The opcode definitions are a little different from 3.12 in that we're going to need to know both the popped and pushed - this is something CPython has tracked internally for a while, it's just we've always just needed the combo. The full list is generated via AI from CPython's list.

Reviewed By: alexmalyshev

Differential Revision: D80752905

fbshipit-source-id: d2d02c92eaafcbd5eeecd20b1feceacead3709ac